fixed bug where you could sever a body part with no sever weapons

+++ Assets/Code/Script/ScriptableObjectsData/WeaponData.cs
@@ -45,5 +45,5 @@ public class WeaponData : ScriptableObject
 
     [Header("Sever")]
     public bool canSever;
-    public float severDmg;
+    public int severDmg;
 }
